updated login redirect URl and organised folder structure
