+++ components/appcontrols/Conferencing.tsx
@@ -11,8 +11,8 @@ import {
 } from '@material-ui/core';
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
 import Switch, {SwitchClassKey, SwitchProps} from '@material-ui/core/Switch';
-import type {FormState} from '../pages/builder';
-import {ProdcuctInfoStyles} from '../styles/ConferencingStyles';
+import type {FormState} from '../../pages/builder';
+import {ProdcuctInfoStyles} from '../../styles/ConferencingStyles';
 interface Conferencing {
   onClickBack: VoidFunction;
   handleValueChange:
